+++ src/apis/questionDetails/dto/create-question-detail.dto.ts
@@ -1,4 +1,5 @@
-import { IsNotEmpty, IsString } from 'class-validator';
+import { IsNotEmpty, IsOptional, IsString } from 'class-validator';
+import { TOPIC, TYPE } from 'src/commons/enum/enum';
 
 export class CreateQuestionDetailDto {
   @IsNotEmpty()
@@ -8,4 +9,16 @@ export class CreateQuestionDetailDto {
   @IsNotEmpty()
   @IsString()
   answer: string;
+
+  @IsOptional()
+  @IsString()
+  library: string;
+
+  @IsNotEmpty()
+  @IsString()
+  topic: TOPIC;
+
+  @IsNotEmpty()
+  @IsString()
+  type: TYPE;
 }
