+++ src/commons/enum/enum.ts
@@ -1,0 +1,21 @@
+export enum TOPIC {
+  JAVA = 'JAVA',
+  JAVASCRIPT = 'JAVASCRIPT',
+  KOTLIN = 'KOTLIN',
+  REACT = 'REACT',
+  NEXT = 'NEXTJS',
+  NEST = 'NESTJS',
+  NODE = 'NODEJS',
+  SPRING = 'SPRING',
+  CS = 'COMPUTER SCIENCE',
+}
+
+export enum TYPE {
+  APP = 'APP DEVELOPMENT', // 모바일 앱 개발
+  WEB = 'WEB DEVELOPMENT', // 웹 개발
+  TOOL = 'DEVELOPMENT TOOL', // 개발 도구
+  LANGUAGE = 'PROGRAMMING LANGUAGE', // 개발 언어
+  SECURITY = 'COMPUTER SECURITY', // 보안
+  DATA = 'DATA MANAGEMENT', // 데이터 관리
+  CAREER = 'DEVELOPER CAREER', // 채용
+}
